QB_2_real_data_loader.py
"""
QB_2: Real Data Loader
Loads actual TCGA/GEO data instead of synthetic data.
"""

import pandas as pd
import numpy as np
from pathlib import Path
from typing import Dict, List, Optional, Tuple
import requests
import gzip
import io
import warnings


class RealTCGALoader:
    """
    Loads real TCGA data from public sources.
    """
    
    def __init__(self, data_dir: Optional[Path] = None):
        self.data_dir = data_dir or Path('data/raw/tcga')
        self.data_dir.mkdir(parents=True, exist_ok=True)
    
    def load_from_gdc(self, project_id: str = 'TCGA-BRCA', 
                     data_category: str = 'Transcriptome Profiling',
                     data_type: str = 'Gene Expression Quantification') -> pd.DataFrame:
        """
        Load TCGA data from GDC (Genomic Data Commons).
        
        Note: This requires GDC API access. For demo, we'll use a sample structure.
        """
        # In production, use GDC API:
        # from gdc import GDCClient
        # client = GDCClient()
        # files = client.get_files(project_id=project_id, ...)
        
        # For now, provide structure for real data loading
        warnings.warn("Using TCGA data structure. Replace with actual GDC API calls.")
        
        # Expected: Download from GDC portal or use gdc-client
        # Files are typically HTSeq-counts or FPKM-UQ format
        
        return None  # Placeholder - implement with actual GDC client
    
    def load_from_file(self, filepath: Path) -> pd.DataFrame:
        """
        Load TCGA expression from local file.
        
        Expected format: genes as rows, samples as columns
        """
        if filepath.suffix == '.gz':
            with gzip.open(filepath, 'rt') as f:
                df = pd.read_csv(f, sep='\t', index_col=0)
        else:
            df = pd.read_csv(filepath, sep='\t', index_col=0)
        
        return df
    
    def get_complexI_genes(self) -> List[str]:
        """
        Get Complex I (NDUF*) gene names.
        """
        # Complex I genes (NADH dehydrogenase)
        complexI_genes = [
            'NDUFV1', 'NDUFV2', 'NDUFV3', 'NDUFS1', 'NDUFS2', 'NDUFS3',
            'NDUFS4', 'NDUFS5', 'NDUFS6', 'NDUFS7', 'NDUFS8', 'NDUFB1',
            'NDUFB2', 'NDUFB3', 'NDUFB4', 'NDUFB5', 'NDUFB6', 'NDUFB7',
            'NDUFB8', 'NDUFB9', 'NDUFB10', 'NDUFB11', 'NDUFA1', 'NDUFA2',
            'NDUFA3', 'NDUFA4', 'NDUFA5', 'NDUFA6', 'NDUFA7', 'NDUFA8',
            'NDUFA9', 'NDUFA10', 'NDUFA11', 'NDUFA12', 'NDUFA13'
        ]
        return complexI_genes
    
    def compute_complexI_score(self, expression: pd.DataFrame) -> pd.Series:
        """
        Compute Complex I score from NDUF* genes (z-scored per gene).
        """
        complexI_genes = self.get_complexI_genes()
        
        # Find available Complex I genes
        available_genes = [g for g in complexI_genes if g in expression.index]
        
        if len(available_genes) == 0:
            # Try alternative naming (e.g., NDUFV1-1, etc.)
            available_genes = [g for g in expression.index if any(cg in g for cg in complexI_genes)]
        
        if len(available_genes) == 0:
            warnings.warn("No Complex I genes found. Using all genes.")
            available_genes = expression.index.tolist()[:20]  # Fallback
        
        # Z-score per gene across samples
        complexI_expr = expression.loc[available_genes]
        complexI_zscored = complexI_expr.apply(lambda x: (x - x.mean()) / x.std(), axis=1)
        
        # Average z-score per sample
        complexI_score = complexI_zscored.mean(axis=0)
        
        return complexI_score


class RealGEOLoader:
    """
    Loads real GEO data.
    """
    
    def __init__(self, data_dir: Optional[Path] = None):
        self.data_dir = data_dir or Path('data/raw/geo')
        self.data_dir.mkdir(parents=True, exist_ok=True)
    
    def load_from_geo(self, gse_id: str) -> Tuple[pd.DataFrame, pd.DataFrame]:
        """
        Load GEO Series data.
        
        Uses GEOquery R package via rpy2 or downloads directly.
        """
        # In production, use:
        # import rpy2.robjects as ro
        # from rpy2.robjects import pandas2ri
        # pandas2ri.activate()
        # ro.r('library(GEOquery)')
        # gse = ro.r(f'getGEO("{gse_id}", GSEMatrix=TRUE)')
        
        warnings.warn(f"GEO loading for {gse_id} - implement with GEOquery or direct download")
        
        return None, None  # Placeholder


class ExpressionProcessor:
    """
    Processes expression data for simulation.
    """
    
    @staticmethod
    def normalize_and_filter(expression: pd.DataFrame,
                            min_expression: float = 0.1,
                            min_samples: int = 3) -> pd.DataFrame:
        """
        Normalize and filter expression data.
        """
        # Filter low expression
        mask = (expression > min_expression).sum(axis=1) >= min_samples
        expr_filtered = expression[mask]
        
        # Log transform (if not already)
        if expr_filtered.max().max() > 100:
            expr_filtered = np.log2(expr_filtered + 1)
        
        # Z-score normalization per gene
        expr_normalized = expr_filtered.apply(
            lambda x: (x - x.mean()) / x.std() if x.std() > 0 else x * 0,
            axis=1
        )
        
        return expr_normalized
    
    @staticmethod
    def split_by_complexI_quartiles(complexI_scores: pd.Series) -> Dict[str, List[str]]:
        """
        Split samples into Low (Q1) vs High (Q4) Complex I quartiles.
        """
        q1 = complexI_scores.quantile(0.25)
        q4 = complexI_scores.quantile(0.75)
        
        low_samples = complexI_scores[complexI_scores <= q1].index.tolist()
        high_samples = complexI_scores[complexI_scores >= q4].index.tolist()
        
        return {
            'Low': low_samples,
            'High': high_samples
        }

