requirements.txt

numpy
scipy
pandas
matplotlib
seaborn
pyyaml
networkx
streamlit
pyarrow
